Ratings/Reviews Completed (almost)
ratings/reviews pages are complete except getAllRatingsByMovieId is not working correctly but once that is complete, finishing up will take a few minutes. Also My Lists page needs work but we can display the same way we do for movie on home page. Also creating a new list creates it twice and no delete method for it that I could see